use svelte as classic frontend
