+++ src/frontend/classic/FrontendInfo.ts
@@ -1,11 +1,9 @@
-import { IFrontendInfo } from '../IFrontend'
+import { IFrontendInfo } from '../IFrontend';
 
 export const Info: IFrontendInfo = {
     ID: 'classic',
     Label: 'Classic',
     Description: 'The standard frontend, no bells no whistles ...',
-    Screenshots: [
-        'https://raw.githubusercontent.com/manga-download/hakuneko/master/assets/screenshot-pages.png'
-    ],
+    Screenshots: [],
     ModuleFile: '/frontend/Classic.js'
 };
