+++ src/frontend/classic/components/Console.svelte
