+++ src/frontend/classic/components/Home.svelte
