+++ src/frontend/classic/components/Jobs.svelte
