+++ src/frontend/classic/components/Media.svelte
