+++ src/frontend/classic/components/MediaItem.svelte
