+++ src/frontend/classic/components/MediaItemSelect.svelte
