+++ src/frontend/classic/components/MediaSelect.svelte
