+++ src/frontend/classic/components/Network.svelte
