+++ src/frontend/classic/components/PluginSelect.svelte
