+++ src/frontend/classic/components/Theme.svelte
