+++ src/frontend/classic/components/Viewer.svelte
