+++ src/frontend/classic/lib/imageLoader.svelte
