+++ src/frontend/classic/lib/smartImageLoader.svelte
