+++ src/frontend/sample-svelte/rollup.config.ts
@@ -1,12 +1,8 @@
+import svelte from 'rollup-plugin-svelte';
 import preprocess from 'svelte-preprocess';
+import resolve from '@rollup/plugin-node-resolve';
 import { terser as minify } from "rollup-plugin-terser";
-import commonjs from "@rollup/plugin-commonjs";
-import copy from 'rollup-plugin-copy';
-import resolve from "@rollup/plugin-node-resolve";
-import svelte from "rollup-plugin-svelte";
 import typescript from '@rollup/plugin-typescript';
-import livereload from "rollup-plugin-livereload";
-import postcss from 'rollup-plugin-postcss';
 
 const isProduction = process.env.ROLLUP_WATCH !== 'true';
 
@@ -27,22 +23,11 @@ export default {
             preprocess: preprocess({ sourceMap: !isProduction })
             //css: css => css.write('css/sample-svelte.css')
         }),
-        postcss({
-            extract: 'css/sample-svelte.css'
-        }),
-        copy({
-            targets: [
-                { src: 'src/frontend/sample-svelte/theme', dest: 'build.web/css' },
-                { src: 'node_modules/carbon-components-svelte/css/all.css', dest: 'build.web/css/theme' },
-            ]
-        }),
         resolve({
             browser: true,
-            dedupe: ["svelte"]
+            dedupe: module => /^svelte(\/|$)/.test(module)
         }),
-        commonjs(),
         typescript(),
-        !isProduction && livereload("build.web"),
         isProduction && minify()
     ]
 };
